scripts/generate-usage-data/requirements.txt
python-dotenv
argparse
requests
uuid
json
random
